Split Helper Classes and Trim Stale Comments
Extract Topology and JobStatus into their own files (one class per
file). Drop stale comments, PR-number references, prior-state stories,
ASCII section dividers, and over-explained docstrings across the
launchers package.
