+++ src/swiss_ai_model_launch/launchers/__init__.py
@@ -1,8 +1,10 @@
 from .firecrest_launcher import FirecRESTLauncher
-from .launch_args import LaunchArgs, Topology
+from .job_status import JobStatus
+from .launch_args import LaunchArgs
 from .launch_request import LaunchRequest
-from .launcher import JobStatus, Launcher
+from .launcher import Launcher
 from .slurm_launcher import SlurmLauncher
+from .topology import Topology
 
 __all__ = [
     "FirecRESTLauncher",
