+++ src/swiss_ai_model_launch/launchers/topology.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class Topology(BaseModel):
+    replicas: int = 1
+    nodes_per_replica: int = 1
